Fixed some errors and cleaned up the code.
- Removed the extra `JFrame` and used `DevitoGenerator` as its own
`JFrame`.
- Removed the `implements ActionListener` from `DevitoGenerator`, since
it didn't have the method it needed.
- Changed `imageLabel` to set its image icon instead of making a new
`JLabel` when the button is pressed.
- Usually, using `DISPOSE_ON_CLOSE` is fine, but all threads must be
stopped for the program to end.  That's why I added a `WindowAdapter`
that disposes the `JFrame` when someone tries to close it but also
cancels the timer.  Otherwise, the timer thread will keep the program
from ending.
- Finally, I moved the images to the project directory.  Doing so means
that you can refer to them using a relative path, not an absolute one.
An absolute path won't work on anyone else's computer.