Fixes to make both the timer and the button press work
- Both the timer and the button call the `updateImage()` method.
- The `alreadyDisplayed` list keeps track of `ImageIcon` objects that
have already been displayed.
- When `updateImage()` is called:
	- It checks to see if `alreadyDisplayed` is the length of
`dannyImageArray` and clears it if it is (since all images have been
shown).
	- Then it randomly selects an image until it finds one that is *not*
present in `alreadyDisplayed`.
	- Finally, it adds that image to `alreadyDisplayed` and sets the icon
of `imageLabel` to the image.
	- Note that `updateImage()` is synchronized in case multiple threads
are accessing it at the same time.
	- This implementation will occasionally display the same image twice if
all the images have been shown and the first new image to be shown
happens to be the last image.
- Also fixed the timer so that it uses a correctly created `TimerTask`
object.